+++ client/src/pages/profiles/index.ts
@@ -1,3 +1,4 @@
 export * from './MyProfile';
+export * from './MyProfileEdit';
 export * from './UserProfile';
 export * from './routes';
